Started frontend and incorporated Map functionality

+++ demo/warthog/src/main/java/com/warthogobserver/observation/Observation.java
@@ -26,6 +26,6 @@ public class Observation {
     List<Animal> animals;
 
     String date;
-    private int lng, lat;
+    private double lng, lat;
 
 }
